Add console logging and exception handling
